requirements.txt
torch==2.1.0
transformers==4.35.2
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.26.2
fastapi==0.104.1
uvicorn==0.24.0.post1
jinja2==3.1.2
motor==3.3.1
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.6
matplotlib==3.8.2
pymongo==4.6.0